Remove JNI method unregistering on AT device closing.
Quoting from dalvik/vm/Jni.c, UnregisterNatives function:

The JNI docs refer to this as a way to reload/relink native libraries,
and say it "should not be used in normal native code".  In particular,
there is no need to do this during shutdown, and you do not need to do
this before redefining a method implementation with RegisterNatives.

Fixes the AT command app crashing when
- selecting portal & tools
- selecting something else
- selecting portal & tools again
- (all that while keeping the USB cable connected)

Change-Id: Ib42e23f390cde3088dae4251b695c7edbcee1397